# Optimize the transformRect and transformPoint methods in matrix_utils…